Have done a lot:
Set up the whole database and migration.

Added User model and added repositories (auth and user) with methods to register users and authenticate them.

Made Commands And Queries for registering and authenticating a user,

added pipeline behaviors (validation and logging)

AuthController with two endpoints (login & register)

Also added tests 2 each, one for successfully signing in and one for failing, and also the same for registering a user.